make it better

+++ src/components/ui/ui-equipos/DeleteEquipoBtn.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import FeatherIcon from 'feather-icons-react';
 import { useDispatch } from 'react-redux';
+import Swal from 'sweetalert2';
+
 import { equipoClearRowSelectedAction, equipoStartDeleteAction } from '../../../actions/equiposActions';
 
 
@@ -10,10 +12,26 @@ export const DeleteEquipoBtn = () => {
     const dispatch = useDispatch();
 
     const handleDelete = () => {
-        console.log('Eliminado');
+        // console.log('Eliminado');
 
-        dispatch( equipoStartDeleteAction() );
-        dispatch( equipoClearRowSelectedAction() );
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "¡No podras revertir la acción!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+          }).then((result) => {
+
+                if (result.isConfirmed) {
+
+                dispatch( equipoStartDeleteAction() );
+                dispatch( equipoClearRowSelectedAction() );
+
+            }
+
+        });
     }
